extracted git commit and arguments models into models folder
