+++ lib/argument_parser.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:path/path.dart' as p;
+import 'models/arguments.dart';
 
 Future<Arguments> parseArguments(List<String> arguments) async {
   final parser = ArgParser()
@@ -116,24 +117,4 @@ Duration frequencyOptionToDuration(String option) {
     default:
       return Duration(days: 365);
   }
-}
-
-class Arguments {
-  final String projectLocation;
-  final File psalmConfig;
-  final DateTime from;
-  final DateTime to;
-  final Duration frequency;
-  final String psalmVersion;
-  final bool considerAllCommits;
-
-  Arguments(
-    this.projectLocation,
-    this.psalmConfig,
-    this.from,
-    this.to,
-    this.frequency,
-    this.psalmVersion,
-    this.considerAllCommits,
-  );
 }
